requirements.txt
# FastAPI and web framework dependencies
fastapi==0.112.*
uvicorn[standard]==0.30.*
pydantic==2.5.0
python-multipart==0.0.20
pydantic-settings==2.1.0
cachetools==6.2.0

# Elastic Search and recommendation model
elasticsearch==8.14.0
sentence-transformers==5.1.0

# HTTP client
yake==0.4.8
httpx==0.27.*

# AI/ML dependencies
pandas==2.3.2
ollama==0.3.3
python-dotenv==1.0.0

# Database
pymongo==4.6.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: Database (uncomment if needed)
# sqlalchemy==2.0.23
# alembic==1.13.0
# asyncpg==0.29.0  # for PostgreSQL
# aiosqlite==0.19.0  # for SQLite
